Move SyncPlay code to plugin directory
