+++ src/plugins/syncPlay/core/Helper.js
