+++ src/plugins/syncPlay/core/Settings.js
