+++ src/plugins/syncPlay/core/players/GenericPlayer.js
