+++ src/plugins/syncPlay/core/players/PlayerFactory.js
