+++ src/plugins/syncPlay/ui/groupSelectionMenu.scss
