+++ src/plugins/syncPlay/ui/players/QueueManager.js
