+++ src/plugins/syncPlay/ui/settings/editor.html
